+++ apps/terrarium/Groupe_H/Embedded/include/Embedded/mock_esp8266.h
@@ -34,10 +34,10 @@ namespace mock_esp8266 {
 	
 	int nAPs = 2;
 
-	int 	mock_open(const char *path, int oflag);
-	int     mock_close(int fd);
-	ssize_t mock_read(int fd, void *buf, size_t nbytes);
-	ssize_t mock_write(int fd, const void *buf, size_t nbytes);
+	int mock_open(const char *path, int oflag);
+	int mock_close(int fd);
+	int mock_read(int fd, void *buf, size_t nbytes);
+	int mock_write(int fd, const void *buf, size_t nbytes);
 
 	void mock_classification(const void *buf, size_t nbytes);
 	void mock_testCmdProcess(std::string instr);
